+++ examples/custom_right_panel.rs
@@ -18,8 +18,8 @@ impl MyApp {
 }
 
 impl eframe::App for MyApp {
-    fn update(&mut self, ctx: &egui::Context, _frame: &mut eframe::Frame) {
-        egui::CentralPanel::default().show(ctx, |ui| {
+    fn ui(&mut self, ui: &mut egui::Ui, _frame: &mut eframe::Frame) {
+        egui::CentralPanel::default().show_inside(ui, |ui| {
             if ui.button("Pick single").clicked() {
                 self.file_dialog.pick_file();
             }
@@ -38,7 +38,7 @@ impl eframe::App for MyApp {
             }
 
             self.file_dialog
-                .update_with_right_panel_ui(ctx, &mut |ui, dia| {
+                .update_with_right_panel_ui(ui, &mut |ui, dia| {
                     if dia.mode() == DialogMode::PickMultiple {
                         ui.heading("Selected items");
                         ui.separator();