+++ examples/pick_directory.rs
@@ -18,15 +18,15 @@ impl MyApp {
 }
 
 impl eframe::App for MyApp {
-    fn update(&mut self, ctx: &egui::Context, _frame: &mut eframe::Frame) {
-        egui::CentralPanel::default().show(ctx, |ui| {
+    fn ui(&mut self, ui: &mut egui::Ui, _frame: &mut eframe::Frame) {
+        egui::CentralPanel::default().show_inside(ui, |ui| {
             if ui.button("Pick directory").clicked() {
                 self.file_dialog.pick_directory();
             }
 
             ui.label(format!("Picked directory: {:?}", self.picked_directory));
 
-            if let Some(path) = self.file_dialog.update(ctx).picked() {
+            if let Some(path) = self.file_dialog.update(ui).picked() {
                 self.picked_directory = Some(path.to_path_buf());
             }
         });