+++ examples/save_file.rs
@@ -18,15 +18,15 @@ impl MyApp {
 }
 
 impl eframe::App for MyApp {
-    fn update(&mut self, ctx: &egui::Context, _frame: &mut eframe::Frame) {
-        egui::CentralPanel::default().show(ctx, |ui| {
+    fn ui(&mut self, ui: &mut egui::Ui, _frame: &mut eframe::Frame) {
+        egui::CentralPanel::default().show_inside(ui, |ui| {
             if ui.button("Save file").clicked() {
                 self.file_dialog.save_file();
             }
 
             ui.label(format!("File to save: {:?}", self.file_path));
 
-            if let Some(path) = self.file_dialog.update(ctx).picked() {
+            if let Some(path) = self.file_dialog.update(ui).picked() {
                 self.file_path = Some(path.to_path_buf());
             }
         });