+++ src/data/directory_content.rs
@@ -1,11 +1,13 @@
-use crate::config::{FileDialogConfig, FileFilter};
 use crate::file_dialog::{SortBy, SortOrder};
-use crate::FileSystem;
-use egui::mutex::Mutex;
 use std::path::{Path, PathBuf};
 use std::sync::{mpsc, Arc};
 use std::time::SystemTime;
 use std::{io, thread};
+
+use egui::mutex::Mutex;
+
+use crate::config::{FileDialogConfig, FileFilter};
+use crate::FileSystem;
 
 #[derive(Clone, Debug)]
 pub struct DirectoryFilter {
@@ -459,7 +461,7 @@ fn load_directory(
         }
 
         if let Some(file_filter) = &filter.file_filter {
-            if entry.is_file() && !(file_filter.filter)(entry.as_path()) {
+            if entry.is_file() && !file_filter.filter.matches(entry.as_path()) {
                 continue;
             }
         }
@@ -498,7 +500,7 @@ fn load_directory(
 /// file icon filters.
 fn gen_path_icon(config: &FileDialogConfig, path: &Path, file_system: &dyn FileSystem) -> String {
     for def in &config.file_icon_filters {
-        if (def.filter)(path) {
+        if def.filter.matches(path) {
             return def.icon.clone();
         }
     }